refactor: move updatestats function to own file
